refactor vector search page tests
